Move a bunch of tests to test_stamp_creator
